pr-generator/requirements.txt
# Requirements for Conjur Policy PR Creator
pyyaml>=6.0
requests>=2.28.0
PyGithub>=1.55
